Small reconfigurement
Changed where some of the files live.
Moved some code into their own files to give the app the appearence that it's organized.
