--- app/index.js
@@ -1,11 +1,0 @@
-
-var myApp = require("./app.js"),
-    server = require('http').Server(myApp);
-    
-var io = require('socket.io')(server);
-
-require("./lib/socketHelper.js")(io);
-
-server.listen(myApp.get("port"), function(){
-    console.log("Server listening on port: %s", myApp.get("port") );
-});
