+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "private": true,
   "scripts": {
-    "start": "node ./app/index.js"
+    "start": "node index.js"
   },
   "description": "A web app",
   "author": {
@@ -13,7 +13,7 @@
   "dependencies": {
     "express": "4.9.0",
     "express-useragent" : "1.0.4",
-    
+
     "body-parser": "1.8.1",
     "cookie-parser": "1.3.3",
 	"ejs": "2.3.4",
